+++ bin/mr-linter
@@ -17,9 +17,9 @@ use ArtARTs36\MergeRequestLinter\Configuration\Resolver\ConfigResolver;
 use ArtARTs36\MergeRequestLinter\Configuration\Resolver\PathResolver;
 use ArtARTs36\MergeRequestLinter\Configuration\Value\EnvTransformer;
 use ArtARTs36\MergeRequestLinter\Configuration\Value\FileTransformer;
-use ArtARTs36\MergeRequestLinter\Console\DumpCommand;
-use ArtARTs36\MergeRequestLinter\Console\InstallCommand;
-use ArtARTs36\MergeRequestLinter\Console\LintCommand;
+use ArtARTs36\MergeRequestLinter\Console\Command\DumpCommand;
+use ArtARTs36\MergeRequestLinter\Console\Command\InstallCommand;
+use ArtARTs36\MergeRequestLinter\Console\Command\LintCommand;
 use ArtARTs36\MergeRequestLinter\Environment\LocalEnvironment;
 use ArtARTs36\MergeRequestLinter\Linter\Runner\RunnerFactory;
 use ArtARTs36\MergeRequestLinter\Rule\DefaultRules;
