+++ src/Console/Command/DumpCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Console;
+namespace ArtARTs36\MergeRequestLinter\Console\Command;
 
 use ArtARTs36\MergeRequestLinter\Contracts\Config\ConfigResolver;
 use Symfony\Component\Console\Command\Command;
