+++ src/Console/Command/HasConfigFileOption.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Console;
+namespace ArtARTs36\MergeRequestLinter\Console\Command;
 
 use ArtARTs36\MergeRequestLinter\Configuration\Resolver\ResolvedConfig;
 use ArtARTs36\MergeRequestLinter\Configuration\User;
